task6

+++ hello_world/6-size.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+
 /**
  * main - Entry point
  *
@@ -6,10 +7,10 @@
  */
 int main(void)
 {
-  printf("Size of char: %u bytes\n",(int) sizeof(char));
-  printf("Size of int: %u bytes\n",(int) sizeof(int));
-  printf("Size of long int: %u bytes\n",(int) sizeof(long int));
-  printf("Size of long long int: %u bytes\n",(int) sizeof(double));
-  printf("Size of float: %u bytes\n",(int) sizeof(float));
-return 0;
+    printf("Size of char: %u bytes\n", (int) sizeof(char));
+    printf("Size of int: %u bytes\n", (int) sizeof(int));
+    printf("Size of long int: %u bytes\n", (int) sizeof(long int));
+    printf("Size of long long int: %u bytes\n", (int) sizeof(long long int));
+    printf("Size of float: %u bytes\n", (int) sizeof(float));
+    return (0);
 }
